feat: added repository layer
- Implemented the repository layer for database interactions.
- Created repository interfaces and classes for entities.
- Added basic CRUD operations.

fix: corrected error in SQL script

- Fixed syntax error in the table creation script.

+++ src/main/java/entity/Match.java
@@ -1,6 +1,5 @@
 package entity;
 
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -17,7 +16,7 @@ public class Match {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "ID")
-    private long id;
+    private Integer id;
 
     @ManyToOne
     @JoinColumn(name = "Player1", referencedColumnName = "ID", table = "Players")
